Turn off some terminal bells.
'set bell-style none' in .inputrc tells Readline to not use the terminal
bell, so there won't be any bells e.g. when there are multiple tab
completions available in the shell or when you hit Backspace too many
times.
The 'less' command reads the LESS environment variable to check for
options, so adding -Q to that variable tells 'less' to not use the
terminal bell at all. This will also affect 'man' and any other command
that uses 'less' as its default pager.