Remove legacy tmux config entries.
Tmux kind of sucks, in the sense that they make breaking changes ALL.
THE. TIME. with no good way for you to dynamically change your config
based on the version being run. I learned recently that my current
way of doing this using BASH comparison wasn't correct, because I was
actually doing lexicographical string comparison. My attempt to switch
to 'awk' (which supports floating point arithmetic and comparison) was
a colossal failure...I apparently couldn't get the if-shell
quoting/escaping right, and how the quoting actually works is
undocumented in the tmux manual (and, worse yet, the behavior changed
somewhere between tmux 1.8 and 2.8, so my attempt wouldn't have worked
anyways...).

So, I guess I'm just going to use the latest tmux all the time now.
Hopefully the tmux devs get to some semblance of stability (and maybe
switch to Semantic Versioning?) at some point in the future to ease this
burden on the users. I'd rather not have to compile and install the
latest tmux on every machine I use it on, but the devs are leaving
me with no choice.